Smooooothed Out the Movement
Alright, this should feel a lot more satisfying to control. Certainly not perfect, but pretty good for Day 2. I still need to add coyote time, though.

I decided to leave the clunky landing mechanic BUT only when the player falls a certain distance. The further the fall, the slower the recovery.