Collision Tweaks
Trying to work out the flaws in the collisions and movement. This involved lots of taking things apart and putting them back together. Most everything works now, but WallGrab still has a couple glitches. There are also a number of error messages that I do not understand (and Google is not helping me with) but they do not seem to affect the game as far as I can tell... yet...